chore: migrate trade to ts
