+++ packages/trader/src/Modules/Trading/index.js
@@ -1,3 +1,3 @@
-import Trade from './Containers/trade.jsx';
+import Trade from './Containers/trade';
 
 export default Trade;
